chore: Rename Nav to CardNav
