+++ src/pages/Pool/index.tsx
@@ -3,7 +3,7 @@ import { ThemeContext } from 'styled-components'
 import { Pair } from '@pancakeswap-libs/sdk'
 import { Button, CardBody } from '@pancakeswap-libs/uikit'
 import { Link } from 'react-router-dom'
-import Nav from 'components/Nav'
+import CardNav from 'components/CardNav'
 import Question from 'components/QuestionHelper'
 import FullPositionCard from 'components/PositionCard'
 import { useUserHasLiquidityInAllTokens } from 'data/V1'
@@ -60,7 +60,7 @@ export default function Pool() {
 
   return (
     <>
-      <Nav activeIndex={1} />
+      <CardNav activeIndex={1} />
       <AppBody>
         <PageHeader title="Liquidity" description="Add liquidity to receive LP tokens">
           <Button id="join-pool-button" as={Link} to="/add/ETH">
